applied tes patch w/ minor fix to alogscan not compiling on older gcc, and ivpcheck not checking if GUI_BUILD option was set. Thanks Toby
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@3355 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/app_alogscan/ScanHandler.cpp
@@ -108,8 +108,8 @@ void ScanHandler::handle(const string& alogfile)
   int start_digits = (int)(log10( m_report.getMaxStartTime())) + 4;
   int stop_digits  = (int)(log10( m_report.getMaxStopTime())) + 4;
 #else 
-  int line_digits  = (int)(log10(m_report.getMaxLines())) + 2;
-  int char_digits  = (int)(log10(m_report.getMaxChars())) + 2;
+  int line_digits  = static_cast<int>(log10(static_cast<double>(m_report.getMaxLines()))) + 2;
+  int char_digits  = static_cast<int>(log10(static_cast<double>(m_report.getMaxChars()))) + 2;
   int start_digits = (int)(log10(m_report.getMaxStartTime())) + 4;
   int stop_digits  = (int)(log10(m_report.getMaxStopTime())) + 4;
 #endif
